+++ build.zig
@@ -98,6 +98,7 @@ pub fn build(b: *std.Build) !void {
     const dep_lz4 = b.dependency("lz4", .{ .target = target, .optimize = optimize });
     const dep_czrex = b.dependency("czrex", .{ .target = target, .optimize = optimize });
     const dep_datetime = b.dependency("datetime", .{ .target = target, .optimize = optimize });
+    const dep_toml = b.dependency("toml", .{ .target = target, .optimize = optimize });
 
     const prebuild_step = b.step("prebuild", "Setup project for build");
 
@@ -125,6 +126,7 @@ pub fn build(b: *std.Build) !void {
     exe.root_module.addImport("luau", dep_luau.module("zig-luau"));
     exe.root_module.addImport("regex", dep_czrex.module("czrex"));
     exe.root_module.addImport("datetime", dep_datetime.module("zdt"));
+    exe.root_module.addImport("toml", dep_toml.module("tomlz"));
 
     b.installArtifact(exe);
 
@@ -158,6 +160,7 @@ pub fn build(b: *std.Build) !void {
     exe_unit_tests.root_module.addImport("luau", dep_luau.module("zig-luau"));
     exe_unit_tests.root_module.addImport("regex", dep_czrex.module("czrex"));
     exe_unit_tests.root_module.addImport("datetime", dep_datetime.module("zdt"));
+    exe_unit_tests.root_module.addImport("toml", dep_toml.module("tomlz"));
 
     const run_exe_unit_tests = b.addRunArtifact(exe_unit_tests);