+++ src/cli.zig
@@ -8,9 +8,11 @@ const CommandMap = Commands.initCommands(&.{
     @import("commands/run.zig").Command,
     @import("commands/test.zig").Command,
     @import("commands/eval.zig").Command,
-    @import("commands/repl/lib.zig").Command,
-
     @import("commands/setup.zig").Command,
+    @import("commands/repl/lib.zig").Command,
+    @import("commands/init.zig").Command,
+
+    @import("commands/luau.zig").Command,
     @import("commands/help.zig").Command,
 
     @import("commands/version.zig").Command,