+++ src/commands/help.zig
@@ -2,8 +2,6 @@ const std = @import("std");
 
 const command = @import("lib.zig");
 
-const zune_info = @import("zune-info");
-
 fn Execute(_: std.mem.Allocator, _: []const []const u8) !void {
     std.debug.print("\x1b[1;3;37mZ\x1b[2;3mune\x1b[0m - A luau runtime\n" ++
         "\n" ++
@@ -14,7 +12,9 @@ fn Execute(_: std.mem.Allocator, _: []const []const u8) !void {
         "  \x1b[1;32mtest     \x1b[0;2m./test.luau      \x1b[0mRun tests in lua/luau file, similar to run.\n" ++
         "  \x1b[1;32msetup    \x1b[0;2m[editor]         \x1b[0mSetup environment for luau-lsp with editor of your choice.\n" ++
         "  \x1b[1;32mrepl                      \x1b[0mStart REPL session.\n" ++
+        "  \x1b[1;32minit                      \x1b[0mCreate initial files & configs for zune.\n" ++
         "\n" ++
+        "  \x1b[1;34mluau     \x1b[0;2m[args...]        \x1b[0mDisplay info from luau.\n" ++
         "  \x1b[1;34mhelp                      \x1b[0mDisplay help message.\n" ++
         "\n" ++
         "\x1b[1mFlags:\x1b[0m\n" ++